splitted test_plugins into a functional file, and unit one
