+++ src/wordpress/tests/test_wordpress.py
@@ -69,15 +69,6 @@ class TestWPConfig:
             wp_default_site_title="My test")
         return WPConfig(wordpress)
 
-    def test_is_installed(self, wp_config):
-        pass
-
-    def test_is_install_valid(self, wp_config):
-        pass
-
-    def test_add_wp_user(self, wp_config):
-        pass
-
 
 class TestWPGenerator:
 
@@ -99,12 +90,3 @@ class TestWPGenerator:
     def test_config(self, wp_generator):
         assert wp_generator.wp_config.installs_locked == settings.DEFAULT_CONFIG_INSTALLS_LOCKED
         assert wp_generator.wp_config.updates_automatic is False
-
-    def test_prepare_db(self, wp_generator):
-        pass
-
-    def test_install_wp(self, wp_generator):
-        pass
-
-    def test_generate(self, wp_generator):
-        assert wp_generator.generate()
